New feature: Align objects to cells of a table
Updated alignment to table features thanks to the input of @Marmootje. Align to row and align to column functions have also been updated. If there are multiple tables in the selection the largest table will be selected (or widest when aligning to rows, or highest when aligning to columns). Also improved error handling.